Massive refactoring to replace the clunky (but faster?) "int[]" representation of algebraic numbers and vectors with OO encapsulations AlgebraicNumber, AlgebraicVector, and AlgebraicMatrix.  Part of the reason for the switch is to use BigRational for all of the terms in an AlgebraicNumber, meaning that we can now support infinite precision.
The new BigRational uses two forms of storage, either a pair of longs, or a
pair of BigIntegers if necessary.  There is an internal function to pre-test
for overflow, to support crossing from the long storage to the BigInteger storage.

Overall, the code is much cleaner and easier to understand with this change,
since "int[]" used to have two possible interpretations.  The code will be easier
to maintain.  In addition, it will be easier to translate to Python or Javascript.

More performance optimization is possible in BigRational and AlgebraicNumber,
in particular by testing for special cases like *1, +1, *0, and +0 in AlgebraicNumber.
Such special cases are already in place in BigRational, for the most part.