Port build to Gradle & decouple Mac specific code.
* Add ./platform folder where platform specific code will be maintained.
* Add ./platform/mac folder where Mac specific code will be maintained.
* Add ./platform/windows folder where Windows specific code will be
maintained. No code is here yet, but the directory structure is
consistent with other peer folders.
* Add ./platform/shared folder where shared code will be maintained.
This should primarily consist of interface definitions, not class
definitions.
* Add ./platform/mac/build.gradle so that the mac folder can be used as
a Gradle subproject when the project is built on a Mac. Use of this
folder as an actual Gradle subproject remains to be clarified, but this
provides a starting point.

* Revert MacAdapter.java to use native code instead of reflection.
* Move ./src/main/java/org/vorthmann/j3d/MacAdapter.java to
./platform/mac/src/main/java/org/vorthmann/extensions/MacAdapter.java
* Move internal Platform.UI interface from
./src/main/java/org/vorthmann/j3d/Platform.java to stand-alone interface
at
./platform/shared/src/main/java/org/vorthmann/interfaces/IHookMacfinder.java
* Rename interface Platform.UI to IHookMacfinder in the MacAdapter and
ApplicationUI classes.

* Update build.gradle, run-vzome.sh and run-vzome.bat to use new Gradle
build directory structure.
* Remove pom.xml which is not used by Gradle but interferes with Gradle
plug-in in NetBeans IDE. Gradle integration with Eclipse is currently
un-tested for this project.
* Update .gitignore to work with Gradle and new directory structure.
Also ignores anything with a .unused extension so that files that are
unused by the project can be kept in local folders without publishing to
github.
* Update README.md to reflect these changes.