Fixed selection order after undo
Deselection removes objects from specific places in the selection, but undo did not reinsert
them at those places.  In the simple case of deselecting everything, undo reversed the selection order,
changing the meaning of cross product, translation, etc.

Now, the side effect records the index at which the object was removed, and reinserts it on redo.

I've added a regression test, using a new version of the CrossProduct edit.