0.11.0 - Symmetry now extends Embedding
Most uses of AlgebraicVector.toRealVector() are now replaced with calls to the Embedding,
either through RenderedModel.renderVector(), or by getting the Embedding from a RenderedManifestation.
Of the remaining uses of AV.toRealVector(), hopefully the only troublesome ones will be those used
to pick a direction from a RealVector, and even there it should not matter too much unless an Embedding
has extreme shear.