Initial attempt at projecting the Gosset polytope to the heptagon field
This implementation generates the correct number of vertices, but it is flawed in
the way it generates edges, producing less than 2000 edges.  Furthermore, most of these
edges are NOT actually edges of the polytope; they just join two vertices that happen to be
reflections of each other in some mirror.