Added ability to scale by different factors
Added AlgebraicField.getNumMultipliers(), which defaults to
getNumIrrationals().

Lots of work on LengthController and NewLengthPanel, of course.  Currently
there is a limitation in that the radio button treatment will probably
fail if getNumMultipliers() > 2; we should switch to a drop-down in that
case.

I also added ControllerTesting, and started to build there a framework
for testing controllers with less coding.

The most impactful change in terms of source diffs was introducing
controller logging.  This required three major changes.  First, to get an
opportunity to log at the right moments, I had to distinguish the Controller
interface (client) calls (actionPerformed, setProperty) from the methods
that implement them in DefaultController and subclasses (doAction and
setModelProperty).  Only the client entry points do the logging.

Second, for the same reason, I had to remove DefaultController.properties()
and two flavors of firePropertyChange().

Finally, so that I could implement DefaultController.getPath(), I had to
finally normalize the way subcontrollers are attached and named, adding
addSubController() and making setNextController() private.

There is still more that can be done.  Rather than logging, we could use
this mechanism for journalling to files, allowing replay or use in tests.