Adding keys to rendered objects
RenderedManifestations now have UUIDs.

The intent is to use the keys/ids to implement deletion, and thus support
an “undo” action sent to the server side.  I haven’t implemented delete
on the client side yet.

Another major change here is to move all the state up to the App.  This is
basically moving me toward Redux.

Another significant change is the addition of batched rendering, as an
experiment to reduce the amount of thrash.  It is not clear that it had
any effect.  In any case, further reading is pointing toward throttled
events (using lodash) and async actions from Redux.