Extracted UndoRedoController
This refactoring establishes a pattern for further cleanup of DocumentController.
It eliminates one layer of “dispatch” code in DocumentModel, and moves
another layer from DocumentController to UndoRedoController.  Finally,
the latter can use a specific model, EditHistory.

I added EditHistory.Listener.publishChanges() to further clean up a little,
but I’m not sure if this is the right pattern for doing SelectionChanges
notifyListeners().