Rationalized the computation of orbit dots for the triangle
Added Symmetry.locateOrbitDots(), and OrbitDotLocator.  The latter calls
AbstractSymmetry.getOrbitTriangle(), which is customized by many symmetries.

Factored out VefVertexExporter, for use below the model level.  It is used
by OrbitDotLocator, to generate a VEF string that can be pasted into vZome,
for debugging the transformation.

OrbitTrianglePanel displays the triangles for all symmetries of all
FieldApplications.