POC handling AdjustSelection* commands with reflection
Simplifying constructors, so there is one standard one that takes just
the EditorModel, and another one that takes more configuration.

The next step will be to eliminate the second constructor, by defining
a better edit lifecycle and using JSON or Properties for configuration,
similar to what is done here with setXmlAttributes.