Mostly got rid of groupInSelection params on edit constructors
The key is the change in ChangeSelection.loadAndPerform().  Since the XmlSaveFormat is
passed in anyway, we can defer the check for groupingDoneInSelection() until here, and remove
the argument from the whole heirarchy of constructors.

I have not removed it everywhere, to keep this commit of a manageable size.  A follow-up
commit will remove the rest.