Enabled QuaternionSymmetry commands
The commands work, but subsequent tools and commands can fail because
the constructions still have 4D vertex data.  For example, "scale up"
does not work.

We need a way to sensibly project to 3D for certain commands, and stay
in 4D for others.  It would be nice to do quaternion rotations on the
4D data, too.  Perhaps this does not make sense as interactive UX, and
should be done in Observable.