More friendly to QuaternionSymmetry
Few users will have access to the H4 symmetry command, etc., but those
who do will have a slightly better experience.

Manifestation values are always 3D, having been projected.  Constructions
can still be 4D or greater, in order to retain the ability to perform
4D symmetries, or to join points in 4D.

Manifestation.toConstruction() is new, and guarantees a 3D construction.
This is now used by ApplyTool, so that scaling, etc. works correctly.
Of course, any tool application will lose 4D data from the new objects.

Manifestation.getFirstConstruction() is new, and is simply a shortcut to
the old iterator .next().

Construction.is3d() is new, and is used in Manifestation.toConstruction()
in order to avoid creating a new construction when the existing ones
are already 3D.

CommandPolygon is now more forgiving when given 4D points, and simply
skips the collinearity and planarity checks.