Command-line file arguments avoid openurl
On Windows, drag-n-drop or double-click results in a file URI being
passed as a command-line argument.  To handle this, we were treating all
command-line path arguments as URLs, and the "openurl" action had code
to handle File URIs.

A better way is to treat paths we *know* are files as files, and use the
"open" action.  "openurl" is reserved for the menu command, and file URIs
are NOT supported there.

Unfortunately, the diffs in ApplicationUI are obscured by reformatting.
Note that we deal with Windows immediately on line 223.  Note also that
we no longer use "defaultAction"... instead we have the if stmt on line 286.