Made Selection, RealizedModel, and Manifestation interfaces
To have any hope of using all the UndoableEdit subclasses in the
browser, via TeaVM compilation to WebAssembly, they must be decoupled
from the current implementations of Selection and RealizedModel.

Even without the WebAssembly goal, this decoupling is a good idea,
enabling the use of all the edits with other implementations of the model
and selection.