Color now attached to Manifestation
Default Manifestation color is still null.  If null, the current coloring rules apply, first
using the rendering style then the symmetry.  Coloring a Manifestation means that
the specific color will always apply.  Hide/show and delete+undo restore colors,
since the colors are stored in the Manifestations in the history.