JOGL rendering shows balls and struts!
First, I got the View3dActivity test working with OpenGlShim, which means
I got RenderingProgram to work with the shim.

Then, I was able to copy some of the logic to JoglScene, and now I can see
balls and struts!

Also, the test class is now moved to desktop, so that the JOGL dependency
can be removed from core (not done yet).

I added a few methods to OpenGlShim as needed by the test.

The aspect ratio is still goofy, but who cares.