Separated lifecycle of shape data from instance data
The instance data in a ShapeClass is now mutable, via replacePositions().

I'm still just doing a wholesale replacePositions() for every addition or
removal of a RenderedManifestation.  It is not clear how to do better than
that.  At least it is only updating one ShapeClass, and only the positions
vertex array.