Optimized rendering
ShapeAndInstances now has a dirty bit, and we defer rebuilding the buffers
until we need to render.  This may mean we cannot get rid of the Animator,
but we could probably add another level of dirty bit.

Now, the green quaternion projection of the omni-truncated 600-cell takes
about two seconds to render initially, where it used to take about 8 with
JOGL.

The current todo list:

* optimize redraw: avoid render if there are no model or camera changes

* optimize shape change: ShapeAndInstances per style

* correct panel lighting

* sort picking hits correctly

* implement transparency

* implement fog

* remove Animator?

* adopt VBOs (vertex buffer objects)

* implement drawNormals, drawOutlines, showFrameLabels, showIcosahedralLabels

* other optimizations?