Support for smarter shape changes
Added RenderingChanges.shapesChanged().  It returns false when unsupported,
and none of the implementations support it yet.

Added RenderedModel.setShapes().  This calls RC.shapesChanged(), and if
it returns false, it falls back to doing setOrbitSource(), as if the
entire symmetry had been changed.  This is obviously overkill, but it has
been the implementation forever.  Now I have laid the groundwork to fix
that.