Implemented JoglScene.shapesChanged()
The real work is in SymmetryRendering.shapesChanged().  It avoids
any impact on the shape vertex arrays (or VBOs, later), just moves instances
from old to new ShapeAndInstances.

This work entailed moving the RenderedManifestation.resetAttributes()
code into RenderedManifestation.

I also had to change the ShapeAndInstances to keep the RMs in a List
rather than a Set, so that the preview strut removal will not "eat" the
"real" rendered objects.