Tuned the view and model trackballs to different speeds
I played around with the idea of an "accelerator", changing the speed
based on the angle since the last mouse event.  No matter what I did,
this caused jerks in the scrolling, since two consecutive event gaps
could be very different.

In the end, I slowed the view trackball for fine control, and still
slowed the main panel trackball a little, for a smoother feel.

The current todo list:

* pick struts and panels

* correct panel lighting

* sort picking hits correctly

* implement transparency

* implement fog

* remove Animator?

* adopt VBOs (vertex buffer objects)

* implement drawNormals, drawOutlines, showFrameLabels, showIcosahedralLabels

* other optimizations?