Fixed image capture for article thumbnails
Handling maxSize the same way we did for Java3d.

What we should probably do here is use this.aspectRatio when this.width
has been set, then we can pass in a very large maxSize and capture images
big enough for Neatoshop, etc.  Actually, for that purpose we would want
minSize.