Fixed the fog implementation
The prior implementation was far too complicated, and was mixing clip
coordinates and world coordinates, probably.

The new implementation works entirely within the fragment shader, and
just applies the fog to the back 40% of the frustum, extracting the depth
from gl_FragCoord.

Also, I removed the Java3d implementation completely.