All file dialogs remember last directory
We store "last-open-directory.*" and "last-save-directory.*" for all file
extensions.  The ".vZome" properties are used as defaults for other
extensions.

These properties are all stored in memory, in ApplicationController
.properties.  There is no persistence.

If we have no saved directory string, the file dialogs will use the system
default behavior, which is usually to open in the last directory used by
the application.