Better file name defaults in export/save dialogs
I'm now calling FileDialog.setFile() to prepare the file name with the
right extension.  If the file has been opened or saved, then the
"window.file" property should be non-null, and I'm using that to set up
the file name, stripping off the extension (presumably .vZome) and adding
the correct one.

Note that the whole business of "window.title", "window.file", and "name"
properties is pretty confusing, in addition to DocumentFrame.mFile state.
I should probably unify "window.file" and "name" properties.  Until this
commit, "window.file" was only set for opened files, not for saved ones.
Now it should be an invariant.