Fix for the wrong-symmetry bug for ReplaceWithShape
In DocumentMenuBar, we look up the *initial* symmetryController, and use
that for many of the items in the "Tools" menu.  This seems bizarre,
since of course it does not adapt to symmetry system changes!

Nonetheless, it may work for most of those commands, but it clearly fails
for ReplaceWithShape, since you can only get the shape for the default
symmetry system.

The fix was easy... use the DocumentController, and there dispatch to
the current symmetryController.

I do need to test some of the "oldTools" like axialsymm... there seems to
be a good chance that they also get stuck in the wrong system.  At the
very least, if they work correctly, it is only because we never even
retain the context from the symmetryController, instead reconstructing
the symmetry system deeper in core.