Fixed the mysterious reversing panel
The purple-strut-model bug that triggered all my efforts at building a
debugger is now resolved.  It turns out that CommandPolygon was not using
the ordered-selection flag, amazingly.  This means that if you build a
panel, then undo just that command, and build the panel again (not redo!),
the second panel will be the reverse of the first.