Add system setting to have zoom level change indicate document is edited
Setting document.trackMagnificationChange=true in the preferences file will cause changes in the camera zoom level to indicate that a document is edited so the user will be prompted to save a document even if just the zoom has changed. Default value is false, so this new behavior will only occur if it is specifically enabled in the preferences file.