Removing AWT dependencies from Controllers, phase 1
I would like to use the Controller layer from within Android, but that
Android does not supply AWT nor allow it.

In this first phase, I have simply removed the ActionEvent argument from
the doAction method.  For the most part, this is a pretty simple change
to all the controllers.  There were two odd cases:

First, ApplicationController needs to forward a few actions to the UI.
I introduced a sub-interface called UI for this purpose, and I also
moved the "browse-..." action handling to the UI, due to the java.awt.Desktop
dependency.

Second, and more painfully, I had to eliminate the PartsPanelActionEvent
that David introduced a while ago.  This wasn't too bad, in the end, just
requiring a little more structure in the action string going to the
controller, which then parses out the parts and sets some props.