Removing AWT dependencies from Controllers, phase 2
Controller no longer implements ActionListener.  Instead, we now have
ControllerActionListener, which is pretty widely used in the UI.

In other places, we simply call the 2-argument Controller.actionPerformed().
In actuality, I don't think we ever use the first argument, the source.
It is not used in DefaultController, and I don't think there are any
overrides.