Application.loadURL() working again
I'm not sure it was really working before, since the core JAR gets loaded
without any resources, thus without any shapes.  Since the fallback shapes
implementation was ScriptedShapes, I was getting a CNFE about some ANTLR
class, though that seems to imply that a resource *was* loaded.  Later I
got an error about loading resources on a null classloader, or something...
very odd.

It could be that the resource removal for JAR plugins is hit-or-miss.

In any case, I have switched the fallback ball shapes to OctahedralShapes,
which is fully algorithmic, with no required data resources, and now the
document load seems to be working.

On the C# side, I'm now using a Job, since it broke rendering when I did
the vZome load during an onClick handler.