Modify FieldApplicationTest to verify the names of all directions
While this is not really an essential test, it does put all of the direction names together in one place for comparison. It will require that any new directions that are added to the app be added to the test as well. It also provides a centralized point to add more detailed direction tests in the future.