Strut/ball grabbing is now perfect
I switched from BoxColliders to MeshColliders.  Note that creating the
MeshColliders programmatically did not work; the initial state is more
correct if I create one on the template object, and just update the mesh
in the script.

I also fixed the dropdown menu initial state, and replaced the big fixed
models with a nicer one, the orange-purple-green snub dodec with lifelike
balls.