Refactored the Adapter class
I'm working toward the ability to send actions, with an instance of the
Adapter for each document.  Since the load Job on the Unity side cannot
pass any references, the load calls are now static.  When the load is
successful, the new Adapter object sends "AdapterReady", at which point
the C# script can call back (statically) to get the Adapter instance.
I haven't implemented that callback yet.

For now, I've verified that all the existing Unity behavior still works.
Next, I'll try an "undo" button, or perhaps "undoAll".  If that works,
I'll move on to sending a "freeMove" command, if I can catch the grab
release event.

I'm deleting the current unity.Controller, since I'm planning to rename
the Adapter class, and it now has the doAction method.