Implemented simple snap-back on grab release
I added the GrabSnapper class, a subclass of OVRGrabbable that calls
ObjectMoved.  For now, I'm not passing any of the current transform,
I'm just sending a "resetRotation" command to the Adapter, which then
sends back a ResetGameObject event.  In the C#, this function simply
restores the object to its original location.

From here, it should be fairly easy to pass down the current location and
rotation, and to snap both to the closest correct values, using a new
UndoableEdit on the Java side.