Position and rotation snapping
This uses the new FreeMove undoable edit.

Added AlgebraicSeries.  The basic concept works well enough, but it is
NOT a substitute for snapping to balls.  Too many balls seem to not lie
on the grid relative to the origin.

Rotation snapping has a bug.  The rotation computation is deriving the
absolute rotation, not the rotation relative to the starting strut
orientation.

I have also discovered that Unity rendering is left-handed.  It seems to
be fine for now, but users will be puzzled if they notice the disparity
between vZome desktop and vZome VR.