Added the FreeMove undoable edit
This edit is needed for VR, where an object can be grabbed and moved to
a new location and orientation.

I also refactored a bit to create Transformation.transform(Construction),
which knows what to do with points, segments, and polygons.

Finally, I deferred the creation of the DapAdapter until we know we need
it.  It was failing to build in Eclipse, missing the DAP JAR for some
reason.