Snapping to balls is better
I added the RenderedModel.getNearbyBall() method, but it is not always
finding the balls it should.

One thing I should try: look for a ball near *either* end of the strut.
It could be that many of the failures are because I'm putting the "far"
end of the strut near a ball.