Moved orbit set defaults to SymmetryPerspective
The motivation was to define default build orbits for dodecagonal symmetry
in the root3 system.  Since dodecagon symmetry is not a 3D symmetry group,
it would not be correct to control this with zone.getRotationPermutation().
The right thing is to put the control in the hands of the
SymmetryPerspective.

Everything is backward compatible; only the RootThreeFieldApplication
overrides the default behavior, and only for dodecagon symmetry, for now.