Adjust the heptagonal antiprism orbit triangle vertices
Changing the length and correcting the sense of one vector so the triangle can be used to make the heptagonal trackball shape instead of a 14-gonal antiprism. 
This doesn't affect its use in calculating the orbit dot positions since that calculation uses a projection onto the orbit triangle which only cares about the direction of these vectors, not their magnitude or sense.