Added a simple main to test ApplicationController
This is so that I can try out the controller with CheerpJ.  The JARs all
seem to compile fine (except for JOGL dependencies that I'm not worrying
about), and the code runs as well, at least until the Gist URL seems to
return empty data.

See the build.bash under cheerpj to test.  It is tested on a Mac; it
assumes you have installed CheerpJ and done the "control-open" trick on
all the binaries to get them to run despite not being signed.