Cheerp test page is now able to parse a local file upload
I'm using a file input and FileReader to select and "upload" a local
vZome file.  I then pass the file in to Java using cheerpjAddStringFile,
then pass the "/str/..." path in for the ApplicationController.main() to
load.

The only real error I encountered in the Java code was the use of Thread
.getContextClassLoader() in a number of places.  I know I've gone back
and forth on this over the years, but my Googling today seems to clearly
indicate that I should *always* use getClass().getClassLoader().  I've
made those changes; it may cause trouble with Oculus or perhaps the server
for client/server, but if so, I'll deal with it by passing ClassLoaders
where necessary.