Integrated cheerpj approach with the existing React client
Fileopener.js is a basic React component I found online, not even migrated
to the functional style.  It is using local state for the selectedFile.

Since the cheerpj runtime is only available as a script full of globals,
I'm using the script tag in index.html, then accessing the functions as
"window.cheerpjInit", "window.cheerpjRunMain", and so on.  I'll want to
improve on that in the future.