Client app is now successfully loading a file and rendering to JSON
I had a ton of headaches with Jackson, making CheerpJ fail at runtime.
I have sent a test project to them on Gitter; hopefully they find and
fix a defect.  Meanwhile, I have eliminated all the use of ObjectWriter
.writeValueAsString and ObjectMapper.valueToTree().  They seem to have
trouble with doubles.

In doing this, I have probably broken the UnityMesh view of getShapeNode.

I'm still just sending the JSON to stdout.  The next step is to create a
"native" Javascript method to call back and trigger the render actions.