Commit message and comment about mapping shape vertices
Incidentally, all my recent changes to the JsonMapper were due to a defect
in CheerpJ.  According to Alessandro Pignotti on https://gitter.im/leaningtech/cheerpj,
Jackson was trying to retrieve Java generics information for primitive types.
This had be special cased to return null.  He provided a build with a fix,
so I could try to revert all those JSON serialization changes.

One of the changes meant that Polyhedron vertices come across as {x,y,z}
rather than [x,y,z].  This triggered a corresponding simplification in
the shape reducer, hard to spot in the diff because it moved to a function.