Improved Redux architecture for client
I implemented some of the practices from Human Redux, and refactored the
FileOpener as a functional component.  I moved all the behavioral code
into action-events, and used redux-thunk to hide the socket behavior
associated with urlProvided and viewClosed events.

Along the way, I cleared the model on FILE_SELECTED.

Except for the performance, this is a usable functional milestone.  In the
desktop version, opening a file does not replay all the rendering, but I
think that is because we send all adds/removes to the Scene, but we don't
have rendering turned on.  Here, we experience all the adds and removes.