Modularized the reducers, middleware, etc
All the behavior is now in the bundles, and I'm using combineReducers
and the vzomejava.middleware now in index.js.  I added that middleware
so that the files bundle could be self-contained and generic; the middleware
acts on the FILE_LOADED event.

I also simplified the event dispatch from Java to Javascript, by constructing
valid typed action events directly on the Java side.

The vzomejava bundle will probably get deconstructed further, as I add
features.  Otherwise, I'm pretty happy with the encapsulation here.