WIP sending camera details on load
The three.js camera is not yet updating after the dispatch, so I'm going
to incorporate PerspectiveCamera from drei and hope that simplifies the
integration.  It seems that react-three-fiber is a little too unopinionated.

Other than that, things are mostly working.  The trackball controls are not yet consistent
with the camera, but at least the orientation and field-of-view look right.