Preload vZome logo as vZome
As soon as the JRE is ready, I'm now opening vZomeLogo.vZome from a resource.

I improved files.fileSelected(), to dispatch the same events as fetchModel.

The rendering and camera changes are deferred until MODEL_LOADED, so we
don't get a blank screen while a model file loads.

Finally, I've corrected the default backgroundColor so it matches the one
that loads.