Added tooltips and initial progress spinner
I would prefer to wait until after the canvas component is loaded, to
avoid the initial gray progress screen, but it is too complicated to
pass the startProgress action through the components.  (I can't put
the useEffect outside of the <Canvas/> element.)

I'll try again after I have solved the Redux / R3F state issue.

Also, the tooltips are just the built-in "title" attributes on the HTML
elements.  These are ugly, but I was not able to get the CSS working
correctly to use this solution:

https://www.w3schools.com/css/css_tooltip.asp