Export DAE is working as a file download
I had a lot of trouble trying to read from the exported file from the
Javascript side, so I just moved the read to the Java side, and pass the
byte array across.  This corresponds to the cheerpj demo approach:

  https://cheerpjdemos.leaningtech.com/iTextDemo.html

Some tweaks to the state are still necessary, to retain the file name.