WIP Eliminating most use of double in favor of float, for JSweet
This work is motivated by my experimentation with JSweet, the Java-to-JS
transpiler.  I still have 68 "internal transpiler error" messages to deal
with.  I've already dealt with all the normal errors that it emitted, by
excluding classes that required ANTLR, and adding Jackson "candies" in the
pom.xml.

Since there was no candy available for javax.vecmath, I incorporated the
source, something I have considered many times in the past.

A number of the errors and internal errors arose from javax.vecmath, and
I know that JSweet cannot deal with type erasure between Double and Float,
so I decided it was a good time to standardize on floats.  This is a pretty
pervasive change.

I have also removed some unused sources from javax.vecmath.

Finally, and importantly, using JSweet required switching to Java 11.  It
is possible that some of my internal errors are due to code that won't
compile for Java 11... I don't have the Gradle build working, so I can't
tell.