JSweet build for RealVector is working
I have stubbed FloatBuffer and NumberFormat, and now the .ts and .js
files are generated correctly.

There are two problems:

1. I cannot direct the output into the client folder, without causing an error.
I can work around this with file moves.

2. The JSweet Gradle plugin does not use its own sourceSet, so I have to
add the src/jsweet/java path to the main sourceSet.  That makes the main
Java build fail.  I have cloned the jsweet-gradle-plugin project, and I'm
trying to make the fix.  In fact, the plugin should not define its own
includes and excludes, but should use the Gradle sourceSet mechanism.  This will
require an upgrade to Java 11, and that requires a gradlew upgrade, so
this is snowballing a bit.