RealVector static init call working
I'm able to redirect just the JS output to the correct client/public
location, and I have a script tag in index.html that runs RealVector
static init correctly... I can set a breakpoint.

Now I'm going to try to get it to work as a module, so I can use it in
bundles.