Introduced HasRenderedObject and DomParser
DomParser is a good separation, splitting the parsing (with all its
dependencies) out from the DOM stuff.

The HasRenderedObject is a bit harder to motivate, but it does let me
separate those few edits that require the RenderedObject, such as
ReplaceShape.  For my current purpose (the NewCentroid edit POC), it
means I can continue to skip all the Symmetry related classes.  (That
will probably be the next reason to sort edits: those that are pure
geometry, and those that need Symmetry.)

All unit tests pass, and the app builds and runs in Eclipse and Gradle.