Factored XmlSymmetryFormat out of XmlSaveFormat
This is another split between pure geometry and symmetry.

Generally, XmlS*Format needs to be seriously rethought.  It uses some
pretty bad patterns.  To start with, the format capabilities need to be
separated from the parsing and serialization code.

I now have a closure of dependencies for NewCentroid, and I'm getting
past the first phase Java errors.  Now I have 161 errors in the next
phase, most of which are due to DOM dependencies.

All unit tests pass, and the desktop app launches.