NewCentroid constructor success
I can successfully call the NewCentroid constructor.  I'm not doing anything
else at the moment, just demonstrating that the whole bundle loads without
error.

The key was to grab the j4ts bundle.js from .jsweet/candies/js, and put
it into the public folder next to the generated bundle, and then including
script tags for both in index.html.

I'd rather use ES6 modules, but it seems that the transpile won't work
correctly with any of the module settings.

Next step: set up the context for the edit!