Success creating an AlgebraicVector
I have switched streams, starting to focus on building up the editing
infrastructure, so I can actually run the NewCentroid command.

I have constructed shims for AF, AN, RealizedModel, Selection, and
EditorModel.  None of them have any JSweet tricks inside yet, being still
mostly stubs.  The AF and AN implementations were implemented (AF as
a copy of PentagonField but with int[]), just enough to construct an
AlgebraicVector and call toRealVector().

My next focus will be to build up the EM, RM and Selection, to the point
where I can perform the edit, still with no connection to the Redux store.

After that, I'll hook up to the store, and perhaps connect the rendering.
Or I could move the code out of init(), and create a real reducer, action
events, and a simple button UI.