Implemented JsAlgebraicField minimum using JSweet untyped access
I now understand how to call into JS using untyped access.

I have only implemented the methods that were implemented already for this
POC, and I have not yet written the JS side of this, so the calls fail
at runtime.

It turns out I can still do all the Java coding in Eclipse, since I'm
pulling in the jsweet-core with Maven, and the JAR is a dependency in
Eclipse.

I had to roll back to jsweet-core 6.1.0; it appears 6.2.0 was compiled
with Java 11, so it gives me the "55 not 52 class file" error during
JSweet transpiling.

Next step: implement the golden field in JS, grabbing Jacob Rus' code from
Observable.