NewCentroid is working completely!
All interactions with the JS Adapter are working correctly, for this
example, though there are probably still issues with the semantics of
ChangeManifestations.manifestConstruction().  Also, JsAlgebraicField is
still doing its own plus(), rather than delegating.

The interaction between Java and Javascript centers around "manifestations"
represented as int[][][]... an array of AlgebraicVectors, each represented
as int[][], an array of AlgebraicNumbers, each represented in trailing-
divisor form as an int[].  Even a JsBall is a singleton array, for
consistency.

This int[][][] is the value space for the Javascript state, which is
captured as three JS Maps: shown, hidden, selected.