Factored interfaces out of EditorModel
ImplicitSymmetryParameters identifies the edits that still use the implicit
symmetry center.  LegacyEditorModel extends that will support for failed
constructions, used by CommandEdit.

SymmetryAware is the more important one, used in a number of "modern" edits.

The JSweet compile closure is extended to support the ShowPoint command,
which requires only the base EditorModel and ImplicitSymmetryParameters.