Removed dependencies on specific AlgebraicFields
I defined AF.scale4dRoots() and AF.doubleFrameVectors() to abstract the
behaviors in B4Group, F4Group, and OctahedralSymmetry.

The means that all of math/symmetry is now in the JSweet closure.  It also
means that I was able to switch to excludes for algebra, construction, and
commands.

I'm still getting 104 of the "thin" (phase 2) errors from JSweet, but they
don't interrupt TS / JS code gen.