Added implementations toggle
I can now switch between the CheerpJ implementation (the vzomejava bundle)
and the JSweet implementation (the mesh bundle).  Both of them export an
instanceSelector function that returns a ready-to-render model of shapes
and instances.  The instances are normalized, with all values embedded to
reals, all vectors as arrays, and a shapeId.

The rendering in modelcanvas-three.js is therefore uniform, and driven by
the shapes and instances, regardless of whether the instances were in the
state or were derived from the mesh.  Mesh instances will be clickable,
and may be selected or not.

The edit menu is enabled only when the mesh implementation is on.