Mostly succeeded with SymmetrySystem initialization
I also moved the resolveShape(), renderableInstance() and
instanceSelector() functions from mesh to jsweet, so resolveShape() can
use the SymmetrySystem (eventually).

There are still some getClassLoader() errors in the console, but they
don't prevent completion of the SymmetrySystem constructor.

I had to do quite a few things to get this far:

- implement more of AF

- make sure the factors array of JsAN is copied before sharing

- made a new JsAF.createVectorFromTDs() method for JsBall to use, since
   the original createVector() needs to support the symmetry inits, which
   pass AN values as arrays of length 2*order, rather than order+1

- stub out PrintWriter and StringWriter

- refactor the AlgebraicMatrix varargs constructor

- disable the "debugger" for OrbitDotLocator

- change Direction.zoneVectors map from AV keys to String keys

- remove an overloaded constructor for AbstractShapes