Getting shapes to work
The resolveShape() function is now running to the point where resetAttributes
is actually trying to load VEF data.  I have injected the default connector
shape VEF, and the VEF is parsing correctly.

To get the parse to work, I had to implement StringTokenizer in terms of
JS string.split().  Working great, after trimming the final "" string due
to trailing whitespace.

I decoupled RenderedManifestation from RenderedModel, and it was even
a simplification of the Java code.  RM now takes an OrbitSource in the
constructor.

I stubbed UUID, though that will have to become a pass-through to JS, I
believe, for shape IDs.

The next step will be to find a way to actually extract the shape and store
it.  Also, I want to set up a mechanism where the shapes get parsed when
injected, in the background, rather than on demand, perhaps, though the
parsing goes fast.