Selection highlight is working!
I'm not dimming the ambient light yet.

Also I have refactored to use thunks (async actions) for all the jsweet
initialization.  I tried to set it up to pre-load strut shapes, but of
course all I would get is a StrutGeometry, not a Polyhedron.  I'll have
to figure out how to install the strut shapes dynamically.

Finally, I fixed the bug in SymmetrySystem where a null color could be
returned.