New approach for resolving instance shapes, etc.
The approach is working for all commands.

See mesh.createRandom(), which now returns a thunk, so that the mesh can
remain pure, and the resolver is chained as another thunk.

The resolver itself is in jsweet, and it injects itself into mesh on init.

Note that createRandom() seems to offer some possibilities of eliminating
boilerplate, but I want to get more experience before I attempt that.